requirements.txt
asgiref==3.8.1
Django==5.0.7
djangorestframework==3.15.2
geographiclib==2.0
geopy==2.4.1
sqlparse==0.5.1
tzdata==2024.1
